Can now automatically generate lumerical scripts for directional coupler
Added two functions that will generate varFDTD and 3D FDTD scripts that are ready to perform a parameter sweep on the directional coupler generated by 'dir_coupler'. To achieve this, the package must now be installed in the AppData/Roaming/KLayout/python directiory, as opposed to the KLayout directory directly in the users folder, due to some issues with how KLayout's python interprets working directory paths. The package was also renamed to CHICkpea. All constants are now lowercase.