Directional coupler arms more customizable
Shallow s-bends for the directional coupler arms can now be specified. Bend radius is no longer a paramter - it is instead calculated from the supplied length and height of the coupler arms. Also added a function 'dir_coupler_ports' for computing the coordinates of the ports of the generated directional coupler. The README and directional coupler example script were updated to reflect these changes.