All rounded paths are now generated as PCells
Before, the 'round_corners' function was used, which will round the vertexes forming a path, but won't preserve information about the original vertexes and doesn't allow the user to modify the original vertexes, bend radius, and points per full circle in the KLayout GUI after the path is generated. The PCell allows all these things.