Fixed bug with moving generated dir'nal couplers in KLayout GUI
The straight segments seemed to move further than the s-bends when the entire directional coupler was selected and moved in the KLayout graphical editor. This was because the two straight segments were two instances of the same cell (though I don't know why that would cause such odd behavior). Now they have two completely different cells, which seems to circumvent this problem.