Minimal implementation of fully parametrized delay spiral
It's had preliminary testing, and should be fully parametric in terms of spacing, number of turns, and on which side of the spiral its terminations lie, as well as straight-line extensions in the x and y directions. In its current state, the parameters available to the user for tuning are a bit crude and require fiddling, and some of them don't quite make sense anymore, but they are sufficient for design.